+++ plugins/provisioners/ansible/provisioner.rb
@@ -25,6 +25,12 @@ module VagrantPlugins
         # but can be enabled via raw_arguments option.
         options << "--connection=ssh"
 
+        # Increase the SSH connection timeout, as the Ansible default value (10 seconds)
+        # is a bit demanding for some overloaded developer boxes. This is particularly 
+        # helpful when additioal virtual networks are configured, as their availability
+        # is not controlled during vagrant boot process.
+        options << "--timeout=30"
+
         # By default we limit by the current machine, but
         # this can be overridden by the `limit` option.
         if config.limit