+++ CHANGELOG.md
@@ -6,6 +6,7 @@ BUG FIXES:
   - guests/funtoo: fix incorrect path in configure networks [GH-4812]
   - plugins/login: allow users to login with a token [GH-5145]
   - provisioners/ansible: fix SSH settings to support more than 5 ssh keys [GH-5017]
+  - provisioners/ansible: increase ansible connection timeout to 30 seconds [GH-5175]
 
 ## 1.7.2 (January 6, 2015)